server/requirements.txt
asyncio
aiohttp
aiohttp_cors
python-socketio
msgpack